+++ src/solutions/00001-000100/00031-00040/index.mjs
@@ -1,9 +1,11 @@
 import * as problem00031 from './00031';
 import * as problem00032 from './00032';
 import * as problem00033 from './00033';
+import * as problem00034 from './00034';
 
 export default {
   problem00031,
   problem00032,
   problem00033,
+  problem00034,
 };